Pull request #485: SDKCF-4309: Improve Referral App Tracking tests with special characters and provide public API for building URL Scheme and Universal Link
Merge in ECO/core-ios-analytics from feature/SDKCF-4309-special-characters to master

Squashed commit of the following:

commit 7927aae64db033a227c33dd710a669220c233134
Author: ts-malik.alayli <ts-malik.alayli@rakuten.com>
Date:   Wed Oct 6 09:31:37 2021 +0200

    Comments update

commit 994b4c19f155592c95ddf54a9401eb1bc038dbaf
Author: ts-malik.alayli <ts-malik.alayli@rakuten.com>
Date:   Tue Oct 5 15:16:31 2021 +0200

    Examples

commit 69b479014ca87f464287c14161291a71c545a084
Author: ts-malik.alayli <ts-malik.alayli@rakuten.com>
Date:   Tue Oct 5 15:09:24 2021 +0200

    Update

commit 3c87605079886014086475db016a7b40b3d16e06
Author: ts-malik.alayli <ts-malik.alayli@rakuten.com>
Date:   Tue Oct 5 12:08:00 2021 +0200

    Update

commit b72b79c6f4d2b05528aec736f2f8c71bb9a5c332
Author: ts-malik.alayli <ts-malik.alayli@rakuten.com>
Date:   Tue Oct 5 11:18:51 2021 +0200

    Public API improvement

commit 6dae21b7d5afffa3e18cd373c42896e1a7ff1290
Author: ts-malik.alayli <ts-malik.alayli@rakuten.com>
Date:   Tue Oct 5 09:16:53 2021 +0200

    Update

commit f2976467cab4043fc76f762ca7768c9b80327733
Author: ts-malik.alayli <ts-malik.alayli@rakuten.com>
Date:   Tue Oct 5 09:00:25 2021 +0200

    Update

commit 780e4d6ead26184a7fde109d36481a20af8696c3
Author: ts-malik.alayli <ts-malik.alayli@rakuten.com>
Date:   Mon Oct 4 23:13:57 2021 +0200

    Update

commit 7c0cb6e5e2d8a5123f3a73791747014cfcaacae0
Author: ts-malik.alayli <ts-malik.alayli@rakuten.com>
Date:   Mon Oct 4 23:05:50 2021 +0200

    Update

commit 89e7b8f7511ad2e2ffce7c4c40174f919872da32
Author: ts-malik.alayli <ts-malik.alayli@rakuten.com>
Date:   Mon Oct 4 21:37:45 2021 +0200

    Update

commit a25ca34472837fd23c23cce8792697b7c54c10bd
Author: ts-malik.alayli <ts-malik.alayli@rakuten.com>
Date:   Mon Oct 4 21:35:28 2021 +0200

    Update

commit 09f0edb137f25f5cd6b00b7aa8d4db13056e37ea
Author: ts-malik.alayli <ts-malik.alayli@rakuten.com>
Date:   Mon Oct 4 21:29:14 2021 +0200

    Update

commit 6a6f8cb5afe1327a90634906bfa66ecf7d343fce
Author: ts-malik.alayli <ts-malik.alayli@rakuten.com>
Date:   Mon Oct 4 21:17:55 2021 +0200

    Update

commit f2e3d9f25f7e4427726f69eb768cf7d382e58854
Author: ts-malik.alayli <ts-malik.alayli@rakuten.com>
Date:   Mon Oct 4 19:58:01 2021 +0200

    SDKCF-4309: Improve Referral App Tracking tests with special characters

+++ RAnalytics/RAT/RAnalyticsRATTracker.swift
@@ -5,7 +5,7 @@ import CoreLocation
 import RDeviceIdentifier
 
 // swiftlint:disable type_name
-public typealias RAnalyticsRATShouldDuplicateEventCompletion = (_ eventName: String, _ duplicateAccId: Int) -> Bool
+public typealias RAnalyticsRATShouldDuplicateEventCompletion = (_ eventName: String, _ duplicateAccId: Int64) -> Bool
 
 /// Concrete implementation of @ref RAnalyticsTracker that sends events to RAT.
 ///
@@ -738,7 +738,7 @@ extension RAnalyticsRATTracker {
     /// - Parameters:
     ///     - accountId: RAT account ID.
     ///     - applicationId: RAT application ID.
-    @objc(addDuplicateAccountWithId:applicationId:) public func addDuplicateAccount(accountId: Int, applicationId: Int) {
+    @objc(addDuplicateAccountWithId:applicationId:) public func addDuplicateAccount(accountId: Int64, applicationId: Int64) {
         duplicateAccounts.insert(RATAccount(accountId: accountId, applicationId: applicationId, disabledEvents: nil))
     }
 }
